Refactor[NotificationManager]: 변경 NotificationManager 이름 변경 및 object
