+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
         android:theme="@style/Theme.Map"
         tools:targetApi="31">
         <service
-            android:name=".MapFirebaseMessagingService"
+            android:name=".Notification.MapFirebaseMessagingService"
             android:enabled="true"
             android:exported="false">
             <intent-filter>
