+++ app/src/main/java/campus/tech/kakao/map/Notification/NotificationManager.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.Notification
 
 import android.app.NotificationChannel
 import android.app.NotificationManager
@@ -6,10 +6,15 @@ import android.app.PendingIntent
 import android.content.Context
 import android.content.Intent
 import androidx.core.app.NotificationCompat
+import campus.tech.kakao.map.R
 import campus.tech.kakao.map.view.SplashScreen
 
-class Notification {
+object NotificationManager {
 	private lateinit var notificationManager: NotificationManager
+	private const val NOTIFICATION_ID = 222222
+	private const val CHANNEL_ID = "main_default_channel"
+	private const val CHANNEL_NAME = "main channelName"
+	private const val CHANNEL_DESCRIPTION = "main channelDescription"
 	private fun createNotificationChannel(context: Context) {
 		val channel = NotificationChannel(
 			CHANNEL_ID,
@@ -24,8 +29,6 @@ class Notification {
 
 
 	fun createNotification(title: String?, body: String?, context: Context){
-
-
 		createNotificationChannel(context)
 		val intent = Intent(context, SplashScreen::class.java).apply {
 			flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
@@ -52,14 +55,5 @@ class Notification {
 			.setAutoCancel(true)
 
 		notificationManager.notify(NOTIFICATION_ID, builder.build())
-	}
-
-
-
-	companion object {
-		private const val NOTIFICATION_ID = 222222
-		private const val CHANNEL_ID = "main_default_channel"
-		private const val CHANNEL_NAME = "main channelName"
-		private const val CHANNEL_DESCRIPTION = "main channelDescription"
 	}
 }
